backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==1.4.52
pydantic==1.10.12
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.0
joblib==1.3.2
numpy==1.26.4
python-dotenv==1.0.0
# pydantic-settings   <- HAPUS/COMMENT OUT baris ini
